minor routing adjustments & error handling
